[feat] 도메인 및 API 수정
