+++ src/main/java/trackers/demo/gallery/domain/repository/CustomProjectRepository.java
@@ -1,11 +1,10 @@
-package trackers.demo.project.domain.repository;
+package trackers.demo.gallery.domain.repository;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
-import org.springframework.stereotype.Repository;
 import trackers.demo.project.domain.Project;
-import trackers.demo.project.dto.request.ReadProjectFilterCondition;
-import trackers.demo.project.dto.request.ReadProjectSearchCondition;
+import trackers.demo.gallery.dto.request.ReadProjectFilterCondition;
+import trackers.demo.gallery.dto.request.ReadProjectSearchCondition;
 
 public interface CustomProjectRepository {
