+++ src/main/java/trackers/demo/gallery/dto/request/ReadProjectFilterCondition.java
@@ -1,4 +1,4 @@
-package trackers.demo.project.dto.request;
+package trackers.demo.gallery.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
@@ -11,8 +11,8 @@ import java.util.List;
 @AllArgsConstructor
 public class ReadProjectFilterCondition {
 
-    @NotNull(message = "모금 여부를 선택해주세요.(default: false)")
-    private final boolean isDonated;
+//    @NotNull(message = "모금 여부를 선택해주세요.(default: false)")
+//    private final boolean isDonated;
 
     @NotNull(message = "프로젝트 대상을 선택하세요")
     private final List<String> targets;
