+++ src/main/java/trackers/demo/gallery/dto/response/ProjectResponse.java
@@ -1,9 +1,10 @@
-package trackers.demo.project.dto.response;
+package trackers.demo.gallery.dto.response;
 
-import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import trackers.demo.project.domain.Project;
+
+import java.time.LocalDate;
 
 import static lombok.AccessLevel.*;
 
@@ -17,18 +18,28 @@ public class ProjectResponse {
 
     private final String projectTitle;
 
-    private final boolean isLike;
+    private final String summary;
+
+    private final String target;
+
+    private final LocalDate endDate;
+
+    private final Boolean islike;
 
     private final Long likeCount;
 
     public static ProjectResponse of(
             final Project project,
+            final String target,
             final Boolean isLike,
             final Long likeCount) {
         return new ProjectResponse(
                 project.getId(),
                 project.getMainImagePath(),
                 project.getProjectTitle(),
+                project.getSummary(),
+                target,
+                project.getEndDate(),
                 isLike,
                 likeCount
         );
