+++ src/main/java/trackers/demo/project/domain/Project.java
@@ -8,8 +8,7 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 import trackers.demo.member.domain.Member;
 import trackers.demo.global.common.entity.BaseEntity;
 import trackers.demo.project.domain.type.CompletedStatusType;
-import trackers.demo.project.domain.type.DonatedStatusType;
-import trackers.demo.project.dto.request.ProjectCreateSecondRequest;
+import trackers.demo.project.dto.request.ProjectCreateBodyRequest;
 import trackers.demo.project.infrastructure.StringListConverter;
 
 import java.time.LocalDate;
@@ -18,7 +17,6 @@ import java.util.List;
 import static jakarta.persistence.EnumType.*;
 import static jakarta.persistence.GenerationType.*;
 import static trackers.demo.project.domain.type.CompletedStatusType.*;
-import static trackers.demo.project.domain.type.DonatedStatusType.*;
 
 @Entity
 @Getter
@@ -34,11 +32,8 @@ public class Project extends BaseEntity {
     @JoinColumn(name = "member_id")
     private Member member;
 
-    @Column(nullable = false ,columnDefinition = "tinyint(0) default 0")
-    private boolean isRecruit;      // 팀원 모집 여부
-
-    @Column(length = 300)
-    private String wantedMember;    // 희망 팀원
+    @Column(nullable = false, length = 20)
+    private String summary;     // 사회 문제 요약
 
     @Column(nullable = false)
     private LocalDate startDate;    // 프로젝트 시작 날짜
@@ -52,11 +47,11 @@ public class Project extends BaseEntity {
     @Column(nullable = false)
     private String mainImagePath;   // 대표 사진
 
-    @Column(length = 200)
+    @Column(length = 180)
     @Convert(converter = StringListConverter.class)
     private List<String> subTitleList;
 
-    @Column(length = 2000)
+    @Column(length = 3000)
     @Convert(converter = StringListConverter.class)
     private List<String> contentList;
 
@@ -66,14 +61,7 @@ public class Project extends BaseEntity {
 
     @Column(nullable = false)
     @Enumerated(value = STRING)
-    private DonatedStatusType donatedStatus;      // 모집 전, 모집 중
-
-    @Column(nullable = false)
-    @Enumerated(value = STRING)
     private CompletedStatusType completedStatus;    // 임시 저장 상태, 완료 상태
-
-    @Column(nullable = false)
-    private int donatedAmount = 0;   // 후원 받은 금액
 
     @Column(name = "is_deleted")
     private boolean deleted;
@@ -81,8 +69,7 @@ public class Project extends BaseEntity {
     public Project(
             final Long id,
             final Member member,
-            final boolean isRecruit,
-            final String wantedMember,
+            final String summary,
             final LocalDate startDate,
             final LocalDate endDate,
             final String projectTitle,
@@ -90,15 +77,12 @@ public class Project extends BaseEntity {
             final List<String> subTitleList,
             final List<String> contentList,
             final List<String> projectImageList,
-            final DonatedStatusType donatedStatus,
             final CompletedStatusType completedStatus,
-            final int donatedAmount,
             final boolean deleted
     ) {
         this.id = id;
         this.member = member;
-        this.isRecruit = isRecruit;
-        this.wantedMember = wantedMember;
+        this.summary = summary;
         this.startDate = startDate;
         this.endDate = endDate;
         this.projectTitle = projectTitle;
@@ -106,16 +90,13 @@ public class Project extends BaseEntity {
         this.subTitleList= subTitleList;
         this.contentList = contentList;
         this.projectImageList = projectImageList;
-        this.donatedStatus = donatedStatus;
         this.completedStatus = completedStatus;
-        this.donatedAmount = donatedAmount;
         this.deleted = deleted;
     }
 
     public static Project of(
             final Member member,
-            final Boolean isRecruit,
-            final String wantedMember,
+            final String summary,
             final LocalDate startDate,
             final LocalDate endDate,
             final String projectTitle,
@@ -124,8 +105,7 @@ public class Project extends BaseEntity {
         return new Project(
                 null,
                 member,
-                isRecruit,
-                wantedMember,
+                summary,
                 startDate,
                 endDate,
                 projectTitle,
@@ -133,20 +113,17 @@ public class Project extends BaseEntity {
                 null,
                 null,
                 null,
-                NOT_DONATED,
                 NOT_COMPLETED,
-                0,
                 false);
     }
 
-    public void createProject(
-            final ProjectCreateSecondRequest createRequest,
+    public void saveProject(
+            final ProjectCreateBodyRequest createRequest,
             final List<String> projectImageList
             ){
         this.subTitleList = createRequest.getSubtitleList();
         this.contentList = createRequest.getContentList();
         this.projectImageList = projectImageList;
-        this.completedStatus = COMPLETED;
     }
