--- src/main/java/trackers/demo/project/domain/ProjectSubject.java
@@ -1,34 +1,0 @@
-package trackers.demo.project.domain;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import trackers.demo.global.common.entity.BaseEntity;
-
-import static jakarta.persistence.FetchType.*;
-import static jakarta.persistence.GenerationType.*;
-import static lombok.AccessLevel.*;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = PROTECTED)
-public class ProjectSubject extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "project_id")
-    private Project project;
-
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "subject_id")
-    private Subject subject;
-
-    public ProjectSubject(Long id, Project project, Subject subject) {
-        this.id = id;
-        this.project = project;
-        this.subject = subject;
-    }
-}
