--- src/main/java/trackers/demo/project/domain/Subject.java
@@ -1,22 +1,0 @@
-package trackers.demo.project.domain;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class Subject {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(length = 10, nullable = false)
-    private String subjectTitle;
-
-}
