--- src/main/java/trackers/demo/project/domain/repository/SubjectRepository.java
@@ -1,11 +1,0 @@
-package trackers.demo.project.domain.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import trackers.demo.project.domain.Subject;
-
-import java.util.Optional;
-
-public interface SubjectRepository extends JpaRepository<Subject, Long> {
-
-    Optional<Subject> findBySubjectTitle(String subjectName);
-}
