+++ src/main/java/trackers/demo/project/domain/type/CompletedStatusType.java
@@ -2,9 +2,11 @@ package trackers.demo.project.domain.type;
 
 public enum CompletedStatusType {
 
-    COMPLETED,
 
-    NOT_COMPLETED;
+    COMPLETED,
+    NOT_COMPLETED,
+
+    CLOSED;
 
     public static CompletedStatusType mappingType(final boolean isCompleted){
         if(isCompleted) return COMPLETED;
