+++ src/test/java/trackers/demo/project/fixture/ProjectFixture.java
@@ -2,7 +2,6 @@ package trackers.demo.project.fixture;
 
 import trackers.demo.project.domain.Project;
 import trackers.demo.project.domain.type.CompletedStatusType;
-import trackers.demo.project.domain.type.DonatedStatusType;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -14,8 +13,7 @@ public class ProjectFixture {
     public static final Project DUMMY_PROJECT = new Project(
             1L,
             MEMBER,
-            true,
-            "열정있는 팀원을 원합니다.",
+            "중장년층 실업 문제",
             LocalDate.of(2024, 7, 1),
             LocalDate.of(2024, 7, 2),
             "은퇴 후 사업 시작 안전하게!",
@@ -23,8 +21,6 @@ public class ProjectFixture {
             List.of("소제목1", "소제목2"),
             List.of("본문1", "본문2"),
             List.of("project-image1.png", "project-image2.png"),
-            DonatedStatusType.NOT_DONATED,
             CompletedStatusType.COMPLETED,
-            100000,
             false);
 }
