+++ src/test/java/trackers/demo/project/presentation/ProjectControllerTest.java
@@ -1,6 +1,5 @@
 package trackers.demo.project.presentation;
 
-import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.Cookie;
 import org.junit.jupiter.api.BeforeEach;
@@ -10,23 +9,15 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.mapping.JpaMetamodelMappingContext;
 import org.springframework.mock.web.MockMultipartFile;
-import org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders;
 import org.springframework.restdocs.payload.JsonFieldType;
-import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.ResultActions;
 import trackers.demo.global.ControllerTest;
 import trackers.demo.loginv2.domain.MemberTokens;
 import trackers.demo.project.domain.type.CompletedStatusType;
-import trackers.demo.project.dto.request.ProjectCreateFirstRequest;
-import trackers.demo.project.dto.request.ProjectCreateSecondRequest;
-import trackers.demo.project.dto.request.ReadProjectFilterCondition;
-import trackers.demo.project.dto.request.ReadProjectSearchCondition;
-import trackers.demo.project.dto.response.ProjectDetailResponse;
-import trackers.demo.project.dto.response.ProjectResponse;
+import trackers.demo.project.dto.request.ProjectCreateOutlineRequest;
+import trackers.demo.project.dto.request.ProjectCreateBodyRequest;
 import trackers.demo.project.service.ImageService;
 import trackers.demo.project.service.ProjectService;
 
@@ -38,7 +29,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.doNothing;
-import static org.mockito.Mockito.when;
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static org.springframework.http.HttpHeaders.LOCATION;
 import static org.springframework.http.HttpMethod.*;
@@ -54,9 +44,6 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static trackers.demo.global.restdocs.RestDocsConfiguration.field;
-import static trackers.demo.project.fixture.ProjectFixture.DUMMY_PROJECT;
-import static trackers.demo.project.fixture.SubjectFixture.DUMMY_SUBJECT;
-import static trackers.demo.project.fixture.TargetFixture.DUMMY_TARGET;
 
 @WebMvcTest(ProjectController.class)
 @MockBean(JpaMetamodelMappingContext.class)
@@ -83,22 +70,11 @@ public class ProjectControllerTest extends ControllerTest {
         given(jwtProvider.getSubject(any())).willReturn("1");
     }
 
-    protected MockMultipartFile getMockMultipartFile() {
-        String name = "projectMainImage";
-        String contentType = "multipart/form-data";
-        String path = "./src/test/resources/static/images/projectMainImage.png";
-
-        return new MockMultipartFile(name, path, contentType, path.getBytes(UTF_8));
-    }
-
-
     // 테스트용 더미 데이터
-    private void makeProjectFirst() throws Exception{
-        final ProjectCreateFirstRequest projectCreateFirstRequest = new ProjectCreateFirstRequest(
+    private void makeProjectOutline() throws Exception{
+        final ProjectCreateOutlineRequest projectCreateFirstRequest = new ProjectCreateOutlineRequest(
                 "아동",
                 "건강한 삶",
-                true,
-                "열정있는 팀원을 원합니다.",
                 LocalDate.of(2024, 6, 25),
                 LocalDate.of(2024, 8, 25),
                 "은퇴 후 사업 시작 안전하게"
@@ -119,37 +95,6 @@ public class ProjectControllerTest extends ControllerTest {
         );
 
         performPostRequest(projectMainImage, createRequest);
-    }
-
-    private void makeProjectSecond() throws Exception{
-
-        final ProjectCreateSecondRequest projectCreateSecondRequest = new ProjectCreateSecondRequest(
-                List.of("소제목1", "소제목2"),
-                List.of("본문1", "본문2")
-        );
-
-        final MockMultipartFile projectImage1 = new MockMultipartFile(
-                "files",
-                "project1.jpg",
-                "multipart/form-data",
-                "./src/test/resources/static/images/project1.jpg".getBytes()
-        );
-
-        final MockMultipartFile projectImage2 = new MockMultipartFile(
-                "files",
-                "project2.png",
-                "multipart/form-data",
-                "./src/test/resources/static/images/project2.png".getBytes()
-        );
-
-        final MockMultipartFile createRequestFile = new MockMultipartFile(
-                "dto",
-                null,
-                "application/json",
-                objectMapper.writeValueAsString(projectCreateSecondRequest).getBytes(UTF_8)
-        );
-
-        performPostRequest(projectImage1, projectImage2, createRequestFile);
     }
 
     private ResultActions performPostRequest(
@@ -181,39 +126,13 @@ public class ProjectControllerTest extends ControllerTest {
                 .cookie(COOKIE));
     }
 
-    private ResultActions performGetRequest(final int projectId) throws Exception{
-        return mockMvc.perform(RestDocumentationRequestBuilders.get("/project/{projectId}", projectId)
-                .header(AUTHORIZATION, MEMBER_TOKENS.getAccessToken())
-                .cookie(COOKIE)
-                .contentType(APPLICATION_JSON));
-    }
-
-    private ResultActions performGetRequest(
-            final Pageable pageable,
-            final ReadProjectSearchCondition searchCondition,
-            final ReadProjectFilterCondition filterCondition
-    ) throws Exception {
-        return mockMvc.perform(RestDocumentationRequestBuilders.get("/project/gallery")
-                .queryParam("page", String.valueOf(pageable.getPageNumber()))
-                .queryParam("size", String.valueOf(pageable.getPageSize()))
-                .queryParam("sortType", "new")
-                .queryParam("title", searchCondition.getTitle())
-                .queryParam("isDonated", String.valueOf(filterCondition.isDonated()))
-                .queryParam("targets", filterCondition.getTargets().toArray(new String[0]))
-                .header(AUTHORIZATION, MEMBER_TOKENS.getAccessToken())
-                .cookie(COOKIE)
-                .contentType(APPLICATION_JSON));
-    }
-
     @DisplayName("프로젝트를 임시 저장(생성)할 수 있다.")
     @Test
-    void createProjectFirst() throws Exception{
+    void saveProjectOutline() throws Exception{
         // given
-        final ProjectCreateFirstRequest projectCreateFirstRequest = new ProjectCreateFirstRequest(
+        final ProjectCreateOutlineRequest projectCreateFirstRequest = new ProjectCreateOutlineRequest(
                 "아동",
                 "건강한 삶",
-                true,
-                "열정있는 팀원을 원합니다.",
                 LocalDate.of(2024, 6, 25),
                 LocalDate.of(2024, 8, 25),
                 "은퇴 후 사업 시작 안전하게"
@@ -292,14 +211,15 @@ public class ProjectControllerTest extends ControllerTest {
 
     @DisplayName("프로젝트 등록을 완료할 수 있다.")
     @Test
-    void createProjectSecond() throws Exception{
+    void saveProjectBody() throws Exception{
         // given
-        makeProjectFirst();
-        doNothing().when(projectService).validateProjectByMember(anyLong(), any(CompletedStatusType.class));
+        makeProjectOutline();
+        doNothing().when(projectService).validateProjectByMemberAndProjectId(anyLong(), anyLong(), any(CompletedStatusType.class));
 
-        final ProjectCreateSecondRequest projectCreateSecondRequest = new ProjectCreateSecondRequest(
+        final ProjectCreateBodyRequest projectCreateSecondRequest = new ProjectCreateBodyRequest(
                 List.of("소제목1", "소제목2"),
-                List.of("본문1", "본문2")
+                List.of("본문1", "본문2"),
+                List.of("태그1", "태그2", "태그3", "태그4", "태그5")
         );
 
         final MockMultipartFile projectImage1 = new MockMultipartFile(
@@ -358,162 +278,4 @@ public class ProjectControllerTest extends ControllerTest {
                 ));
     }
 
-    @DisplayName("ProjectId로 단일 프로젝트를 조회한다.")
-    @Test
-    void getProject() throws Exception{
-        // given
-        makeProjectFirst();
-        makeProjectSecond();
-        doNothing().when(projectService).validateProjectByProjectId(anyLong());
-        when(projectService.getProjectDetail(any(), any()))
-                .thenReturn(ProjectDetailResponse.projectDetail(
-                        DUMMY_PROJECT,
-                        DUMMY_TARGET,
-                        DUMMY_SUBJECT,
-                        false,
-                        0L));
-
-        // when
-        final ResultActions resultActions = performGetRequest(1);
-
-        // then
-        final MvcResult mvcResult = resultActions.andExpect(status().isOk())
-                .andDo(restDocs.document(
-                        pathParameters(
-                                parameterWithName("projectId")
-                                        .description("프로젝트 ID")
-                        ),
-                        responseFields(
-                                fieldWithPath("projectId")
-                                        .type(JsonFieldType.NUMBER)
-                                        .description("프로젝트 ID")
-                                        .attributes(field("constraint", "양의 정수")),
-                                fieldWithPath("projectTitle")
-                                        .type(JsonFieldType.STRING)
-                                        .description("프로젝트명")
-                                        .attributes(field("constraint", "문자열")),
-                                fieldWithPath("projectTarget")
-                                        .type(JsonFieldType.STRING)
-                                        .description("프로젝트 대상")
-                                        .attributes(field("constraint", "문자열")),
-                                fieldWithPath("projectSubject")
-                                        .type(JsonFieldType.STRING)
-                                        .description("프로젝트 주제")
-                                        .attributes(field("constraint", "문자열")),
-                                fieldWithPath("mainImagePath")
-                                        .type(JsonFieldType.STRING)
-                                        .description("프로젝트 대표 이미지")
-                                        .attributes(field("constraint", "이미지 경로")),
-                                fieldWithPath("subTitleList")
-                                        .type(JsonFieldType.ARRAY)
-                                        .description("소제목 리스트")
-                                        .attributes(key("constraint").value("1개 이상의 문자열(최대 100자)")),
-                                fieldWithPath("contentList")
-                                        .type(JsonFieldType.ARRAY)
-                                        .description("본문 리스트")
-                                        .attributes(key("constraint").value("1개 이상의 문자열(최대 2000자)")),
-                                fieldWithPath("projectImageList")
-                                        .type(JsonFieldType.ARRAY)
-                                        .description("프로젝트 사진 리스트")
-                                        .attributes(key("constraint").value("최대 10장의 사진 파일")),
-                                fieldWithPath("wantedMember")
-                                        .type(JsonFieldType.STRING)
-                                        .description("희망 팀원")
-                                        .attributes(field("constraint", "300자 이하의 문자열")),
-                                fieldWithPath("donatedStatus")
-                                        .type(JsonFieldType.STRING)
-                                        .description("프로젝트 후원 여부")
-                                        .attributes(field("constraint", "문자열")),
-                                fieldWithPath("donatedAmount")
-                                        .type(JsonFieldType.NUMBER)
-                                        .description("후원 금액")
-                                        .attributes(field("constraint", "양의 정수")),
-                                fieldWithPath("likeCount")
-                                        .type(JsonFieldType.NUMBER)
-                                        .description("좋아요 수")
-                                        .attributes(field("constraint", "양의 정수")),
-                                fieldWithPath("like")
-                                        .type(JsonFieldType.BOOLEAN)
-                                        .description("좋아요 여부")
-                                        .attributes(field("constraint", "True: 좋아요 반영, False: 좋아요 해제"))
-                        )
-                )).andReturn();
-
-        final ProjectDetailResponse response = objectMapper.readValue(
-                mvcResult.getResponse().getContentAsString(),
-                ProjectDetailResponse.class
-        );
-        assertThat(response).usingRecursiveComparison()
-                .isEqualTo(ProjectDetailResponse.projectDetail(
-                        DUMMY_PROJECT,
-                        DUMMY_TARGET,
-                        DUMMY_SUBJECT,
-                        false,
-                        0L));
-
-    }
-
-    @DisplayName("조건에 알맞는 프로젝트를 모두 조회할 수 있다")
-    @Test
-    void getAllProjectsByCondition() throws Exception{
-        // given
-        makeProjectFirst();
-        makeProjectSecond();
-        when(projectService.getAllProjectsByCondition(
-                any(), any(), any(), any()))
-                .thenReturn(List.of(ProjectResponse.of(DUMMY_PROJECT, false, 0L)));
-
-        ReadProjectSearchCondition searchCondition = new ReadProjectSearchCondition("");
-        ReadProjectFilterCondition filterCondition = new ReadProjectFilterCondition(true, List.of("실버 세대", "청소년"));
-        Pageable pageable = PageRequest.of(1, 5);
-
-        // when
-        final ResultActions resultActions = performGetRequest(pageable, searchCondition, filterCondition);
-
-        // then
-        final MvcResult mvcResult = resultActions.andExpect(status().isOk())
-                .andDo(restDocs.document(
-                        queryParameters(
-                                parameterWithName("page").description("페이지 번호 (1부터 시작)"),
-                                parameterWithName("size").description("한 페이지에 프로젝트 개수 (default: 5)"),
-                                parameterWithName("sortType").description("정렬 타입: new(default), likeCount(좋아요 순), recentTime(최신 순), closingTime(종료임박 순)"),
-                                parameterWithName("title").description("프로젝트 제목에 포함된 단어 검색"),
-                                parameterWithName("isDonated").description("모금 여부 (true/false)"),
-                                parameterWithName("targets").description("프로젝트 대상 리스트 (null일 때, 전체 대상 검색)")
-                        ),
-                        responseFields(
-                                fieldWithPath("[].projectId")
-                                        .type(JsonFieldType.NUMBER)
-                                        .description("프로젝트 ID")
-                                        .attributes(field("constraint", "양의 정수")),
-                                fieldWithPath("[].mainImagePath")
-                                        .type(JsonFieldType.STRING)
-                                        .description("프로젝트 대표 이미지")
-                                        .attributes(field("constraint", "이미지 경로")),
-                                fieldWithPath("[].projectTitle")
-                                        .type(JsonFieldType.STRING)
-                                        .description("프로젝트명")
-                                        .attributes(field("constraint", "문자열")),
-                                fieldWithPath("[].likeCount")
-                                        .type(JsonFieldType.NUMBER)
-                                        .description("좋아요 수")
-                                        .attributes(field("constraint", "양의 정수")),
-                                fieldWithPath("[].like")
-                                        .type(JsonFieldType.BOOLEAN)
-                                        .description("좋아요 여부")
-                                        .attributes(field("constraint", "True: 좋아요 반영, False: 좋아요 해제"))
-                        )
-                ))
-                .andReturn();
-
-        final List<ProjectResponse> projectResponses = objectMapper.readValue(
-                mvcResult.getResponse().getContentAsString(),
-
-                new TypeReference<>() {
-                }
-        );
-        assertThat(projectResponses).usingRecursiveComparison()
-                .isEqualTo(List.of(ProjectResponse.of(DUMMY_PROJECT, false, 0L)));
-
-    }
 }
